explorer/requirements.txt
numpy
trimesh
viser
fastapi
uvicorn[standard]
jinja2
httpx
Pillow
